[Gold III] Title: 별자리 만들기, Time: 0 ms, Memory: 2692 KB -BaekjoonHub

+++ 백준/Gold/4386. 별자리 만들기/별자리 만들기.cc
@@ -4,7 +4,7 @@ using namespace std;
 #define Y second
 
 vector<pair<double, double>> coord;
-vector<tuple<double, double, double>> adj;
+vector<tuple<double, int, int>> adj;
 vector<int> p(100005, -1);
 
 int find(int x) {
